+++ contracts/Interfaces/IRiskFund.sol
@@ -7,7 +7,15 @@ pragma solidity 0.8.13;
  * @notice Interface implemented by `RiskFund`.
  */
 interface IRiskFund {
-    function transferReserveForAuction(address comptroller, address bidder, uint256 amount) external returns (uint256);
+    function transferReserveForAuction(
+        address comptroller,
+        address bidder,
+        uint256 amount
+    ) external returns (uint256);
 
-    function updatePoolState(address comptroller, address asset, uint256 amount) external;
+    function updatePoolState(
+        address comptroller,
+        address asset,
+        uint256 amount
+    ) external;
 }